Update custom activities documentation
This commit modifies the custom activities topic, updating the language of the introduction. The changes aim to provide a clearer description of Elsa's prebuilt activities and introduce the concept of creating domain-specific, custom activities. It also enhances the forward to give a more detailed overview of what the guide will cover.